% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmcForwardBackward}
\alias{ctmcForwardBackward}
\title{Forward-backward Algorithm for Symmetric Continuous-time HMM (Rcpp)}
\usage{
ctmcForwardBackward(ts, theta, obs)
}
\arguments{
\item{ts}{size m positive float values characterize the timestamps for the observations}

\item{theta}{a single float value as the tansition parameter determining the transition probabilities}

\item{obs}{size n*m matrix containing all the conditional probabilities for each observation given each state at that timestamp.}
}
\value{
m*n matrix containing all the conditional probabilities of each state at the timestamps given the observations
}
\description{
This function implements the forward-backward algorithm for the continuous-time Hidden Markov Model in a specific condition: n uniform initial states and a special-case dual transition probabilities related to the timestamp and parameter theta (detailed formula in the pdf instruction). The input arguments include the timestamps of each observation represented in a row with size m, a transition parameter theta used for the definition of the transition probabilities, a matrix containing all the conditional probabilities for each observation given each state at that timestamp. Note the time complexity is O(mn).
}
\examples{
obs <- matrix(c(0.88,0.10,0.88,0.10,0.02,0.30,0.02,0.30,0.10,0.60),2,5)
theta <- log(2)
ts <- c(1,2,3,4,5)
ctmcForwardBackward(ts, theta, obs)

ts <- c(1,2.95,3,4,5)
ctmcForwardBackward(ts, theta, obs)

}
